requirements-test.txt
aiohttp==3.9.3
aiosignal==1.3.1
astroid==3.1.0
attrs==23.2.0
black==24.2.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coverage[toml]==7.4.3
cryptography==42.0.5
dill==0.3.8
distlib==0.3.8
filelock==3.13.1
fixinventoryclient==2.0.0
flake8==7.0.0
frozenlist==1.4.1
hypothesis==6.99.5
idna==3.6
iniconfig==2.0.0
isort==5.13.2
jsons==1.6.3
mccabe==0.7.0
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pep8-naming==0.13.3
pip==24.0
pip-tools==7.4.1
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyjwt==2.8.0
pylint==3.1.0
pyproject-api==1.6.1
pyproject-hooks==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.5.post1
pytest-cov==4.1.0
setuptools==69.2.0
sortedcontainers==2.4.0
toml==0.10.2
tomlkit==0.12.4
tox==4.14.1
types-aiofiles==23.2.0.20240311
typing-extensions==4.10.0
typish==1.9.3
virtualenv==20.25.1
wheel==0.43.0
yarl==1.9.4
